requirements.txt
opencv-python
python-chess
keyboard
numpy
stockfish
pyautogui
mss
python-dotenv
pywin32